Move planEntry definition and helper functions to own file and add ability to check for plan delivery
